machine learning/requirements.txt
# Core libraries
numpy
pandas
dask
ray

# Modeling
xgboost_ray
lightgbm>=3.3.0
# torch>=2.0.0
scikit-learn>=1.2.0

# MLOps
mlflow>=2.3.0
optuna>=3.0.0
joblib>=1.2.0
cloudpickle>=2.2.0
azure-identity
azure-ai-ml

# API
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0

# Dashboarding / Visualization
plotly>=5.13.0
dash>=2.10.0
streamlit>=1.22.0

# Utilities
python-dotenv>=1.0.0

# Optional but useful
pyyaml
hydra-core>=1.3.0
great_expectations>=0.17.20
prefect>=2.10.0

# For local MLflow UI (optional)
gunicorn